+++ GEON-PPANG-iOS/Global/Network/DataModel/Common/BreadResponseType.swift
@@ -1,0 +1,26 @@
+//
+//  BreadResponseType.swift
+//  GEON-PPANG-iOS
+//
+//  Created by 이성민 on 2023/07/19.
+//
+
+import Foundation
+
+struct BreadResponseType: Codable, Hashable {
+    let breadTypeId: Int
+    let breadTypeName: String
+    let isGlutenFree: Bool
+    let isVegan: Bool
+    let isNutFree: Bool
+    let isSugarFree: Bool
+    
+    func configureTrueOptions() -> [(String, Bool)] {
+        var optionsBoolArray: [(String, Bool)] = []
+        if isGlutenFree { optionsBoolArray.append((I18N.BakeryList.glutenfree, true)) }
+        if isVegan { optionsBoolArray.append((I18N.BakeryList.vegan, true)) }
+        if isNutFree { optionsBoolArray.append((I18N.BakeryList.nutfree, true)) }
+        if isSugarFree { optionsBoolArray.append((I18N.BakeryList.noSugar, true)) }
+        return optionsBoolArray
+    }
+}
