+++ GEON-PPANG-iOS/Global/Network/DataModel/Filter/FilterRequestDTO.swift
@@ -25,32 +25,3 @@ struct FilterRequestDTO: Codable {
                                   isNotOpen: false)
     }
 }
-
-struct BreadRequestType: Codable {
-    var isGlutenFree: Bool
-    var isVegan: Bool
-    var isNutFree: Bool
-    var isSugarFree: Bool
-    
-    func isNoneSelected() -> Bool {
-        if isGlutenFree == false && isVegan == false && isNutFree == false && isSugarFree == false {
-            return true
-        } else {
-            return false
-        }
-    }
-}
-
-struct NutrientRequestType: Codable {
-    var isNutrientOpen: Bool
-    var isIngredientOpen: Bool
-    var isNotOpen: Bool
-    
-    func isNoneSelected() -> Bool {
-        if isNutrientOpen == false && isIngredientOpen == false && isNotOpen == false {
-            return true
-        } else {
-            return false
-        }
-    }
-}
