+++ GEON-PPANG-iOS/Presentation/BakeryList/ViewControllers/Model/BakeryListResponseDTO.swift
@@ -21,23 +21,3 @@ struct BakeryListResponseDTO: Hashable, BakeryListProtocol {
     let isBookMarked: Bool
     let bookMarkCount: Int
 }
-
-// MARK: - BreadType
-
-struct BreadResponseType: Codable, Hashable {
-    let breadTypeId: Int
-    let breadTypeName: String
-    let isGlutenFree: Bool
-    let isVegan: Bool
-    let isNutFree: Bool
-    let isSugarFree: Bool
-    
-    func configureTrueOptions() -> [(String, Bool)] {
-        var optionsBoolArray: [(String, Bool)] = []
-        if isGlutenFree { optionsBoolArray.append((I18N.BakeryList.glutenfree, true)) }
-        if isVegan { optionsBoolArray.append((I18N.BakeryList.vegan, true)) }
-        if isNutFree { optionsBoolArray.append((I18N.BakeryList.nutfree, true)) }
-        if isSugarFree { optionsBoolArray.append((I18N.BakeryList.noSugar, true)) }
-        return optionsBoolArray
-    }
-}
